1. Fix an error when fetch string type field may cause malform packet error.    This is beacuse of an const MAX_PHYSICAL_PACKET_LENGTH  in fe should be 2^24 -1,    but it is set as 2^24 -2 by mistake. 2. Fix bitmap_to_string may failed when the result is large than 2G

+++ be/src/exprs/anyval_util.h
@@ -312,7 +312,7 @@ public:
         }
     }
 
-    static StringVal from_buffer(FunctionContext* ctx, const char* ptr, int len) {
+    static StringVal from_buffer(FunctionContext* ctx, const char* ptr, int64_t len) {
         StringVal result(ctx, len);
         memcpy(result.ptr, ptr, len);
         return result;
@@ -323,7 +323,7 @@ public:
         return val;
     }
 
-    static StringVal from_buffer_temp(FunctionContext* ctx, const char* ptr, int len) {
+    static StringVal from_buffer_temp(FunctionContext* ctx, const char* ptr, int64_t len) {
         StringVal result = StringVal::create_temp_string_val(ctx, len);
         memcpy(result.ptr, ptr, len);
         return result;
